Matvey feature

+++ src/main/java/ru/skypro/homework/dto/AdDto.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class AdDto {
-    private Integer author; //id автора объявления
+    private Long author; //id автора объявления
 
     private String image;//ссылка на картинку объявления !
 
-    private Integer pk; //id объявления
+    private Long pk; //id объявления
 
     private Integer price;//цена объявления !
